Translated Pokémon nickname prompt
The original prompt was "<POKEMON>'s NICKNAME?", which is difficult to
translate because it almost necessarily involves putting the Pokémon's
name into the genitive case, which I don't really want to try to do
programatically. Thus, I'm just ignoring the "'s" relationship between
the species name, and the word "nickname".

Перевела промпт для ввода клички Покемона.